fix: rename files and directories according to guidelines (B21)
