+++ src/components/Layout/Layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Header from '../Header/Header';
+import Header from '../header/header';
 
 
 export default function Layout() {
