+++ src/components/offers-list/offers-list.tsx
@@ -1,4 +1,4 @@
-import OfferCard from '../OfferCard/OfferCard';
+import OfferCard from '../offer-card/offer-card';
 import { Offer } from '../../const';
 
 type OffersListProps = {
