+++ src/pages/login/login.tsx
@@ -1,4 +1,4 @@
-import Logo from '../../components/Logo/Logo';
+import Logo from '../../components/logo/logo';
 
 export default function Login() {
   return (
