+++ src/pages/main/main.tsx
@@ -1,5 +1,5 @@
-import OffersFilter from '../../components/OffersFilter/OffersFilter';
-import OffersList from '../../components/OffersList/OffersList';
+import OffersFilter from '../../components/offers-filter/offers-filter';
+import OffersList from '../../components/offers-list/offers-list';
 import { Offer } from '../../const';
 
 type MainProps = {
